Implement the feature to switch to MOP when dependencies change
