+++ emop-maven-plugin/src/main/java/edu/cornell/AffectedSpecsMojo.java
@@ -78,6 +78,10 @@ public class AffectedSpecsMojo extends ImpactedClassMojo {
 
     public void execute() throws MojoExecutionException {
         super.execute();
+        if (dependencyChangeDetected) {
+            getLog().info("Switched to MOP run");
+            return;
+        }
         if (getImpacted().isEmpty()) {
             return;
         }
